use TextureSlice everywhere

+++ src/text.rs
@@ -142,14 +142,14 @@ impl SpriteMaker {
 
                 sprites.push(spright::Sprite {
                     src: spright::Rect {
-                        x: allocation.rectangle.min.x as f32,
-                        y: allocation.rectangle.min.y as f32,
-                        width: allocation.rectangle.width() as f32,
-                        height: allocation.rectangle.height() as f32,
+                        x: allocation.rectangle.min.x,
+                        y: allocation.rectangle.min.y,
+                        width: allocation.rectangle.width() as u32,
+                        height: allocation.rectangle.height() as u32,
                     },
                     dest_size: spright::Size {
-                        width: allocation.rectangle.width() as f32,
-                        height: allocation.rectangle.height() as f32,
+                        width: allocation.rectangle.width() as u32,
+                        height: allocation.rectangle.height() as u32,
                     },
                     transform: spright::AffineTransform::translation(
                         physical_glyph.x as f32 + image.placement.left as f32,
